no `Glyph` size only 128 bit

+++ res/Shaders/text.vert
@@ -1,5 +1,5 @@
 layout(location = 0) in vec2 a_Offset;     // instance world offset
-layout(location = 1) in vec4 a_TexRect;    // instance (uMin, vMin, uMax, vMax)
+layout(location = 1) in uvec4 a_TexRect;    // instance (uMin, vMin, uMax, vMax)
 
 out vec2 v_TexCoord;
